requirements.txt
colorama==0.4.6
google_api_python_client==2.113.0
python-dotenv==1.0.1
python_dateutil==2.8.2
yt_dlp==2023.12.30
